[MICONCALL-323] Refactor state management.
State is refactored into a behaviour, and there are two implementations
of the behaviour available. Ets, which is the original implementation of
assertion statemanagement with an ETS table (this does not work with mulitple
nodes), and Conn, which stores the entire assertion in the Conn (session).
Conn allows for state to be shared accross multiple nodes, which is what we
need.